requirements.txt
bcrypt
boto3
Flask
Flask-Injector
Flask-Pydantic
Flask-JWT-Extended
